将选项卡内容分散不同的页面文件
