+++ VR33B.UI/GraphicGridPage.xaml
@@ -1,0 +1,36 @@
+﻿<Page x:Class="VR33B.UI.GraphicGridPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:VR33B.UI"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="GraphicGridPage">
+
+    <Grid x:Name="GraphicGrid" Margin="10">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition x:Name="GraphicImageColumn" Width="3*"></ColumnDefinition>
+            <ColumnDefinition x:Name="SeparateColumn" Width="auto"></ColumnDefinition>
+            <ColumnDefinition x:Name="GraphicDataColumn" Width="*"></ColumnDefinition>
+        </Grid.ColumnDefinitions>
+        <Grid x:Name="GraphicImageGrid" Grid.Column="0">
+
+        </Grid>
+        <Separator Grid.Column="1" BorderThickness="1" Width="1" Style="{StaticResource {x:Static ToolBar.SeparatorStyleKey}}"></Separator>
+        <Grid x:Name="GraphicDataGrid" Grid.Column="2">
+            <Grid.RowDefinitions>
+                <RowDefinition></RowDefinition>
+                <RowDefinition></RowDefinition>
+            </Grid.RowDefinitions>
+            <StackPanel Grid.Row="0">
+                <TextBlock Text="数据" Style="{DynamicResource MaterialDesignTitleTextBlock}"></TextBlock>
+                <ListBox x:Name="DataListBox"></ListBox>
+            </StackPanel>
+            <StackPanel Grid.Row="1" Margin="0, 10, 0, 0">
+                <TextBlock Text="碰撞时间" Style="{DynamicResource MaterialDesignTitleTextBlock}"></TextBlock>
+                <ListBox x:Name="ImpactListBox"></ListBox>
+            </StackPanel>
+        </Grid>
+    </Grid>
+</Page>
