+++ VR33B.UI/MainWindow.xaml.cs
@@ -25,21 +25,8 @@ namespace VR33B.UI
         public MainWindow()
         {
             InitializeComponent();
-            serialPortNames = new ObservableCollection<string>(SerialPort.GetPortNames());
-            if (serialPortNames.Count == 0)
-            {
-                serialPortNames.Add("无串口");
-            }
-            SerialNoBox.ItemsSource = serialPortNames;
-            SerialNoBox.SelectedItem = serialPortNames[0];
-            BaudRateBox.ItemsSource = baudRates;
-            BaudRateBox.SelectedItem = baudRates[0];
-            DataBitBox.ItemsSource = dataBits;
-            DataBitBox.SelectedItem = dataBits[0];
-            StopBitBox.ItemsSource = Enum.GetValues(typeof(StopBits));
-            StopBitBox.SelectedItem = StopBits.One;
-            ParityBitBox.ItemsSource = Enum.GetValues(typeof(Parity));
-            ParityBitBox.SelectedItem = Parity.None;
+
+
             serialPort = new SerialPort();
             serialPort.DataReceived += SerialPort_DataReceived;
         }
@@ -49,7 +36,7 @@ namespace VR33B.UI
         {
             SerialPort sp = (SerialPort)sender;
             string indata = sp.ReadExisting();
-            ReceivedRawDataBox.AppendText(indata);
+            //ReceivedRawDataBox.AppendText(indata);
         }
 
         private ObservableCollection<string> serialPortNames;
@@ -59,11 +46,11 @@ namespace VR33B.UI
         SerialPort serialPort;
         private void SwitchPortButton_Click(object sender, RoutedEventArgs e)
         {
-            serialPort.PortName = (string)SerialNoBox.SelectedItem;
-            serialPort.BaudRate = (int)BaudRateBox.SelectedItem;
-            serialPort.DataBits = (int)DataBitBox.SelectedItem;
-            serialPort.Parity = (Parity)ParityBitBox.SelectedItem;
-            serialPort.StopBits = (StopBits)StopBitBox.SelectedItem;
+            //serialPort.PortName = (string)SerialNoBox.SelectedItem;
+            //serialPort.BaudRate = (int)BaudRateBox.SelectedItem;
+            //serialPort.DataBits = (int)DataBitBox.SelectedItem;
+            //serialPort.Parity = (Parity)ParityBitBox.SelectedItem;
+            //serialPort.StopBits = (StopBits)StopBitBox.SelectedItem;
             try
             {
                 if (!serialPort.IsOpen)
@@ -81,17 +68,6 @@ namespace VR33B.UI
             }
         }
 
-        private void SamplingThresholdSlider_ValueChanged(object sender, RoutedPropertyChangedEventArgs<double> e)
-        {
-            SamplingThresholdValueBlock.Text = ((int)e.NewValue).ToString() + "%";
-        }
 
-        protected override void OnRenderSizeChanged(SizeChangedInfo sizeInfo)
-        {
-            ReceivedRawDataBox.Height = RowDataGrid.ActualHeight / 2 - 20 - ReceivedRawDataTitleBlock.ActualHeight - 10;
-            SentRawDataBox.Height = RowDataGrid.ActualHeight / 2 - 20 - SentRawDataTitleBlock.ActualHeight - 10;
-            System.Diagnostics.Debug.WriteLine(ReceivedRawDataBox.Height);
-            base.OnRenderSizeChanged(sizeInfo);
-        }
     }
 }
