+++ VR33B.UI/SerialPortControlPage.xaml
@@ -1,0 +1,61 @@
+﻿<Page x:Class="VR33B.UI.SerialPortControlPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:VR33B.UI"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="SerialPortControlPage">
+
+    <Grid Margin="10">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+            <ColumnDefinition Width="auto"></ColumnDefinition>
+            <ColumnDefinition></ColumnDefinition>
+        </Grid.ColumnDefinitions>
+        <StackPanel x:Name="SerialConfigurePanel" Grid.Column="0" VerticalAlignment="Stretch">
+            <StackPanel x:Name="SerialNoPanel" Orientation="Horizontal">
+                <TextBlock Text="串口号：" Style="{DynamicResource MaterialDesignSubheadingTextBlock}" VerticalAlignment="Center"></TextBlock>
+                <ComboBox x:Name="SerialNoBox" Width="60" >
+                </ComboBox>
+            </StackPanel>
+            <StackPanel x:Name="BaudRatePanel" Orientation="Horizontal">
+                <TextBlock Text="波特率：" Style="{DynamicResource MaterialDesignSubheadingTextBlock}" VerticalAlignment="Center"></TextBlock>
+                <ComboBox x:Name="BaudRateBox" Width="60"></ComboBox>
+            </StackPanel>
+            <StackPanel x:Name="DataBitPanel" Orientation="Horizontal">
+                <TextBlock Text="数据位：" Style="{DynamicResource MaterialDesignSubheadingTextBlock}" VerticalAlignment="Center"></TextBlock>
+                <ComboBox x:Name="DataBitBox" Width="60"></ComboBox>
+            </StackPanel>
+            <StackPanel x:Name="StopBitPanel" Orientation="Horizontal">
+                <TextBlock Text="停止位：" Style="{DynamicResource MaterialDesignSubheadingTextBlock}" VerticalAlignment="Center"></TextBlock>
+                <ComboBox x:Name="StopBitBox" Width="60"></ComboBox>
+            </StackPanel>
+            <StackPanel x:Name="ParityBitPanel" Orientation="Horizontal" Margin="0, 0, 0, 10">
+                <TextBlock Text="校验位：" Style="{DynamicResource MaterialDesignSubheadingTextBlock}" VerticalAlignment="Center"></TextBlock>
+                <ComboBox x:Name="ParityBitBox" Width="60"></ComboBox>
+            </StackPanel>
+            <Button x:Name="SwitchPortButton" VerticalAlignment="Bottom" Content="打开/关闭串口" Click="SwitchPortButton_Click"></Button>
+        </StackPanel>
+        <Separator Grid.Column="1" BorderThickness="1" Style="{StaticResource {x:Static ToolBar.SeparatorStyleKey}}"></Separator>
+        <Grid x:Name="RowDataGrid" Grid.Column="2" Margin="10">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="*"></RowDefinition>
+                <RowDefinition Height="*"></RowDefinition>
+            </Grid.RowDefinitions>
+            <StackPanel x:Name="ReceivedRawDataPanel" Grid.Row="0"
+                                    VerticalAlignment="Stretch" Margin="0, 0, 0, 10">
+                <TextBlock x:Name="ReceivedRawDataTitleBlock" Text="原始接收数据" Style="{DynamicResource MaterialDesignTitleTextBlock}" Margin="0, 0, 0, 10"></TextBlock>
+                <RichTextBox x:Name="ReceivedRawDataBox" VerticalAlignment="Stretch" BorderThickness="1" VerticalScrollBarVisibility="Auto" 
+                                         IsReadOnly="True"
+                                         ></RichTextBox>
+            </StackPanel>
+            <StackPanel x:Name="SentRawDataPanel" Grid.Row="1"
+                                    VerticalAlignment="Stretch">
+                <TextBlock x:Name="SentRawDataTitleBlock" Text="原始发送数据" Style="{DynamicResource MaterialDesignTitleTextBlock}" Margin="0, 0, 0, 10" ></TextBlock>
+                <RichTextBox x:Name="SentRawDataBox" VerticalAlignment="Stretch" BorderThickness="1" VerticalScrollBarVisibility="Auto" IsReadOnly="True"></RichTextBox>
+            </StackPanel>
+        </Grid>
+    </Grid>
+</Page>
